backend/requirements.txt
celery
Django
django-celery-beat
django-celery-results
django-cors-headers
django-environ
django-extensions
djangorestframework
google-api-python-client
google-auth-oauthlib
gunicorn
# psycopg2
# Pillow

#dev
coverage
